wear bolus activity and refactoring
