+++ wear/src/main/java/info/nightscout/androidaps/data/BgWatchData.java
@@ -1,4 +1,4 @@
-package info.nightscout.androidaps;
+package info.nightscout.androidaps.data;
 
 /**
  * Created by emmablack on 1/7/15.
