+++ wear/src/main/java/info/nightscout/androidaps/watchfaces/BaseWatchFace.java
@@ -1,4 +1,4 @@
-package info.nightscout.androidaps;
+package info.nightscout.androidaps.watchfaces;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -34,6 +34,11 @@ import com.ustwo.clockwise.common.WatchShape;
 import java.util.ArrayList;
 import java.util.Date;
 
+import info.nightscout.androidaps.data.BasalWatchData;
+import info.nightscout.androidaps.data.BgWatchData;
+import info.nightscout.androidaps.data.ListenerService;
+import info.nightscout.androidaps.R;
+import info.nightscout.androidaps.data.TempWatchData;
 import lecho.lib.hellocharts.view.LineChartView;
 
 /**
