+++ wear/src/main/res/layout/rect_activity_bigchart.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" tools:context=".Home" tools:deviceIds="wear_square"
+    android:layout_height="match_parent" tools:context=".watchfaces.Home" tools:deviceIds="wear_square"
     android:background="@color/black"
     android:id="@+id/main_layout">
