+++ wear/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <string name="app_name">AAPS Prefs.</string>
-    <string name="label_actions_activity">AAPS Actions</string>
+    <string name="app_name">AAPS</string>
+    <string name="label_actions_activity">AAPS</string>
 
     <string-array name="chart_timeframe">
         <item>1 hour</item>
